missed one with the floats

+++ ChapeauDAL/Readers/MenuReader.cs
@@ -25,8 +25,8 @@ namespace ChapeauDAL.Readers
                 menuId: (int)reader["menuId"],
                 menuTypeId: (int)reader["menuTypeId"],
                 name: (string)reader["itemName"],
-                VATRate: (float)reader["VATRate"],
-                price: (float)reader["price"]
+                VATRate: (double)reader["VATRate"],
+                price: (double)reader["price"]
             );
 
             return menuItem;
